REFACTOR: evaluation and move generation code for improved type consistency
- Changed return type of popLSB function from uint8_t to int for better compatibility with move indices.
- Updated all instances of popLSB usage to reflect the new return type.
- Replaced uint8_t types with int in move generation functions to enhance type safety and avoid potential overflows.
- Removed unused evaluation constants related to undefended and attacked pawns, as well as early king penalties.
- Cleaned up UCI interface by removing references to obsolete evaluation constants.
- Adjusted JSON tuning files to remove parameters that are no longer used in the evaluation.

+++ engine/eval/evaluator.hpp
@@ -216,7 +216,7 @@ private:
     static inline int32_t evalCentralBlockPenalty(uint8_t blockerType, int fullMoves) noexcept;
     static inline int32_t evalBlockedPawnByBishopsSide(const chess::Board& b, int side, int fullMoves) noexcept;
     static inline int32_t evalBlockedPawnByBishopsPawn(const chess::Board& b, int side, uint64_t bishops, int fullMoves, int psq) noexcept;
-    static inline uint8_t popLSB(uint64_t& bb) noexcept;
+    static inline int popLSB(uint64_t& bb) noexcept;
 #ifdef DEBUG
     static void traceTerm(int32_t& eval, int32_t delta, const char* label) noexcept;
 #endif
